avance HU

+++ src/main/java/com/riwi/MealMap/dtos/request/Ingredient/IngredientsWithoutId.java
@@ -1,6 +1,5 @@
 package com.riwi.MealMap.dtos.request.Ingredient;
 
-import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -17,7 +16,7 @@ public class IngredientsWithoutId {
     private String name;
 
     @NotNull
-    private Float price;
+    private Double price;
 
     @NotNull
     private String weight;
